yt/requirements.txt
flet==0.22.*
mysql.connector==2.2.*